sketch: fix logic for item/time aggregations
Add a nil to the start of the itemtime aggregations array so that the
indices match up with the M array and later jstar calculations.

This also removes the dumb search for a non-nil time aggregation -- all
the array elements are non-nil, we just don't add any more if we're not
supposed to.